Analyzer service skeleton inside app/, using env variable.
