+++ app/bin/tools/backfill_packages.dart
@@ -6,8 +6,8 @@ import 'dart:async';
 
 import 'package:gcloud/db.dart';
 
-import 'package:pub_dartlang_org/models.dart';
-import 'package:pub_dartlang_org/service_utils.dart';
+import 'package:pub_dartlang_org/frontend/models.dart';
+import 'package:pub_dartlang_org/frontend/service_utils.dart';
 
 Future main(List<String> args) async {
   int updatedPackages = 0;
