+++ app/bin/tools/remove_package.dart
@@ -9,11 +9,11 @@ import 'package:gcloud/db.dart';
 import 'package:gcloud/storage.dart';
 import 'package:pub_semver/pub_semver.dart';
 
-import 'package:pub_dartlang_org/backend.dart';
-import 'package:pub_dartlang_org/models.dart';
+import 'package:pub_dartlang_org/shared/configuration.dart';
 
-import 'package:pub_dartlang_org/configuration.dart';
-import 'package:pub_dartlang_org/service_utils.dart';
+import 'package:pub_dartlang_org/frontend/backend.dart';
+import 'package:pub_dartlang_org/frontend/models.dart';
+import 'package:pub_dartlang_org/frontend/service_utils.dart';
 
 Future main(List<String> arguments) async {
   if (arguments.length < 2 ||
