+++ app/lib/analyzer/api_handlers.dart
@@ -1,0 +1,7 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// TODO: implement API handlers:
+// - get analysis data
+// - trigger analysis
