+++ app/lib/frontend/backend.dart
@@ -14,11 +14,12 @@ import 'package:logging/logging.dart';
 import 'package:pub_server/repository.dart';
 import 'package:uuid/uuid.dart';
 
+import '../shared/package_memcache.dart';
+import '../shared/utils.dart';
+
 import 'model_properties.dart';
 import 'models.dart' as models;
-import 'package_memcache.dart';
 import 'upload_signer_service.dart';
-import 'utils.dart';
 
 final Logger _logger = new Logger('pub.cloud_repository');
