+++ app/lib/frontend/search_service.dart
