+++ app/lib/shared/configuration.dart
