+++ app/test/frontend/backend_test.dart
@@ -12,9 +12,9 @@ import 'package:pub_server/repository.dart' as pub_server;
 import 'package:test/test.dart';
 import 'package:yaml/yaml.dart';
 
-import 'package:pub_dartlang_org/backend.dart';
-import 'package:pub_dartlang_org/models.dart';
-import 'package:pub_dartlang_org/upload_signer_service.dart';
+import 'package:pub_dartlang_org/frontend/backend.dart';
+import 'package:pub_dartlang_org/frontend/models.dart';
+import 'package:pub_dartlang_org/frontend/upload_signer_service.dart';
 
 import 'backend_test_utils.dart';
 import 'utils.dart';
