+++ app/test/frontend/golden/authorized_page.html
