+++ app/test/frontend/golden/pagination_middle.html
