+++ app/test/frontend/golden/pagination_single.html
