+++ app/test/frontend/golden/pkg_index_page.html
