+++ app/test/frontend/golden/search_page.html
